requirements.txt
crc16==0.1.1
poetry==1.1.13
ujson==5.2.0
requests==2.27.1
tvm_valuetypes==0.0.9
pynacl==1.5.0